move some dataset files
