+++ scripts/test/makeDir.sh
@@ -1,0 +1,7 @@
+cd $1
+i=0
+while ((i < $2))
+do
+    mkdir database$i
+    ((i++))
+done
