+++ scripts/test/network_monitor/monitor.sh
@@ -1,3 +1,8 @@
-sudo iptables -nvxt filter -L INPUT | head -n 12 | tail -n 10 > record
-sudo iptables -nvxt filter -L OUTPUT | head -n 12 | tail -n 10 >> record
+#!/bin/bash
+
+set -ex
+
+iptables -nvxt filter -L INPUT | head -n 12 | tail -n 10 > record
+iptables -nvxt filter -L OUTPUT | head -n 12 | tail -n 10 >> record
+echo "Communication Cost: "
 awk '{sum += $2};END {print sum}' record
