+++ scripts/test/network_monitor/start.sh
@@ -1,15 +1,14 @@
 start() {
-  sudo iptables -I INPUT -p tcp --dport $1
-  sudo iptables -I OUTPUT -p tcp --sport $1
+  iptables -I INPUT -p tcp --dport $1
+  iptables -I OUTPUT -p tcp --sport $1
 }
 
-start 12345
-start 12346
-start 12347
-start 12348
-start 12349
-start 12350
-start 12351
-start 12352
-start 12353
-start 12354
+if [ ! -n "$1" ];then
+  echo "Port cannot be empty!"
+  echo "You can use the following command to monitor the port: sudo bash start.sh 8888"
+  exit
+fi
+
+export PORT=$1
+echo "Monitoring port: "  ${PORT}
+start ${PORT}
